+++ config/routes.rb
@@ -6,6 +6,10 @@ RecordingStudioTermsAndConditions::Engine.routes.draw do
     resources :terms, only: %i[index new create show edit update] do
       resources :users, only: :index, controller: "term_users"
     end
+    get "people/:actor_type/:actor_id",
+        to: "people#show",
+        as: :person,
+        constraints: { actor_type: /[A-Za-z][A-Za-z0-9:]*/, actor_id: %r{[^/]+} }
   end
 
   root "acceptances#show"